Flytter mapping av arbeidssokerperiode
Ønsker ikke at RaaData-klassen skal dependes på fra domenet. Dette er en Entity/DTO for databasen, og bør derfor behandles på samme måte som andre DTOer for adaptere.

Flytter da logikk ut fra constuctor og over til Repository (egen mapper), og sørger samtidig for å sortere listen bare én gang. Det gjør koden litt lettere å lese.
